+++ src/common/utils.ts
@@ -1,0 +1,8 @@
+export function isValidJson(str: string) {
+  try {
+    JSON.parse(str)
+  } catch (e) {
+    return false
+  }
+  return true
+}