+++ src/renderer/src/routes/QueueView.svelte
@@ -5,10 +5,11 @@
   import { JsonView } from '@zerodevx/svelte-json-view'
   import { queues } from '../stores/queues'
   import { onDestroy, onMount } from 'svelte'
-  import AddQueueButton from '../components/AddQueueButton.svelte'
-  import { Trash, Icon, BarsArrowUp, BarsArrowDown, Clipboard } from 'svelte-hero-icons'
-  import { selectedExchange, selectedMessage, selectedQueue } from '../stores/ui'
-  import type { Message, Queue } from '@common/types'
+
+  import { Icon, BarsArrowUp, BarsArrowDown, Clipboard } from 'svelte-hero-icons'
+  import { selectedExchange, selectedMessage, selectedQueue, unreadMessageQueues } from '../stores/ui'
+  import QueueTabs from '../components/queue-tabs/QueueTabs.svelte'
+  import type { Message } from '@common/types'
 
 
   const api = window.api
@@ -31,17 +32,22 @@
     selectedMessage.set(null)
   }
 
-  async function deleteQueue(queue) {
-    await api.rabbit.deleteQueue(queue)
-    $queues = $queues.filter((q) => q !== queue)
-  }
-
   onMount(async () => {
     api.rabbit.onMessage((incoming: Message) => {
       // Only add the message if it belongs to the selected queue
-      if (incoming.queueId == $selectedQueue.id) {
+      if ($selectedQueue && incoming.queueId == $selectedQueue.id) {
         messages.update((value) => {
           return [incoming, ...value]
+        })
+      } else {
+        // If it doesn't belong to the selected queue, add the queue to the list of queues with
+        // unread messages
+        const incomingQueue = $queues.find((q) => q.id === incoming.queueId)
+        unreadMessageQueues.update((value) => {
+          if (value.some((v) => v.exchange === incomingQueue.exchange && v.queue === incomingQueue.id)) {
+            return value
+          }
+          return [...value, { exchange: incomingQueue.exchange, queue: incomingQueue.id }]
         })
       }
     })
@@ -54,12 +60,6 @@
   onDestroy(() => {
     api.rabbit.removeMessageListener()
   })
-
-  async function selectQueue(queue: Queue) {
-    $selectedQueue = queue
-    // load messages from main
-    messages.set(await api.rabbit.getMessages(queue.id) || [])
-  }
 
   let sortDirection = 'asc'
 
@@ -78,29 +78,7 @@
   <!-- message section -->
   <div class="h-full w-4/5 bg-primary-800 flex flex-col">
     {#if $selectedExchange !== null}
-      <div class="w-full h-8 bg-primary-600 overflow-x-scroll flex">
-        {#each $queues.filter((q) => q.exchange === $selectedExchange.name) as queue}
-          <button
-            class="w-64 h-full bg-primary-700 hover:bg-primary-500 hover:text-primary-50 hover:font-medium transition-all"
-            on:click={() => selectQueue(queue)}
-            class:selected-queue={$selectedQueue === queue}
-          >
-            <div class="flex items-center justify-between px-2 h-full group">
-              <div class="text-primary-200 text-xs font-light">{queue.name}</div>
-              <button
-                on:click={() => deleteQueue(queue)}
-                class="text-primary-200 text-xs font-extralight ml-2 group-hover:visible invisible transition-opacity"
-              >
-                <Icon src={Trash} class="w-4 h-4" solid />
-              </button
-              >
-            </div>
-          </button>
-        {/each}
-        {#if $selectedExchange}
-          <AddQueueButton />
-        {/if}
-      </div>
+      <QueueTabs />
 
       <!-- message list -->
       <div class="w-full flex bg-primary-950 text-primary-50 text-xs font-light items-center">
@@ -146,7 +124,11 @@
               <div
                 class="group h-full w-full overflow-y-scroll whitespace-pre-wrap font-mono text-primary-100 json-wrap"
               >
-                <JsonView json={$selectedMessage.body} />
+                {#if $selectedMessage.body instanceof Object}
+                  <JsonView json={$selectedMessage.body} />
+                {:else}
+                  {$selectedMessage.body}
+                {/if}
                 <button
                   on:click="{() => copy(JSON.stringify($selectedMessage.body, null, 2))}"
                   class="absolute bottom-5 right-5 transition-opacity group-hover:opacity-75 opacity-0 duration-500 ease-in hover:scale-110">
@@ -183,10 +165,6 @@
 <style lang="postcss">
   .selected {
     @apply bg-primary-500 text-primary-50 font-medium transition-all;
-  }
-
-  .selected-queue {
-    @apply bg-primary-800 text-primary-50 font-medium transition-all;
   }
 
   .fade-right-side {